+++ src/main/java/com/pm/papermanagement/controller/PaperController.java
@@ -11,11 +11,20 @@ import com.pm.papermanagement.mapper.PaperMapper;
 import lombok.Value;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.shiro.SecurityUtils;
+import org.apache.tomcat.util.file.ConfigurationSource;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.web.ServerProperties;
+import org.springframework.core.io.FileSystemResource;
+import org.springframework.core.io.PathResource;
+import org.springframework.core.io.Resource;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -26,6 +35,7 @@ import java.util.UUID;
 
 @Slf4j
 @RestController
+//@CrossOrigin
 public class PaperController {
     PaperMapper paperMapper;
 
@@ -44,14 +54,16 @@ public class PaperController {
         if(ObjectUtils.isEmpty(paperParam.getTitle())){
             throw new BizException("Title不能为空");
         }
-        Integer.parseInt(paperParam.getLibrary_id());
-        paperMapper.addPaper(paperParam);
+        int library_id = Integer.parseInt(paperParam.getLibrary_id());
+        String owner = SecurityUtils.getSubject().getPrincipal().toString();
+        paperMapper.addPaper(library_id,paperParam.getTitle(),owner);
         Paper paper = paperMapper.selectPaperByTitle(paperParam.getTitle());
         return ReturnValue.generateSuccess(new PaperParamOutput(paper));
     }
 
     @GetMapping("/api/paper")
     public ReturnValue requestPaper(@RequestParam(value = "library_id") String library_id, @RequestParam(value = "page_num") int page_num, @RequestParam(value = "page_size") int page_size, @RequestParam(value = "title",required = false) String title){
+        page_num -= 1;
         if(page_num < 0 || page_size < 0){
             throw new BizException("参数不能为负数");
         }
@@ -61,36 +73,92 @@ public class PaperController {
         }else{
             papers = paperMapper.selectPaperBySequence(Integer.parseInt(library_id),page_num * page_size, page_size);
         }
-        return ReturnValue.generateSuccess(new Papers(papers));
+        int paperNum = paperMapper.getPaperNum(Integer.parseInt(library_id));
+        return ReturnValue.generateSuccess(new Papers(papers,paperNum));
     }
 
     @RequestMapping(value = "/api/paper/{paper_id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ReturnValue updatePaper(@PathVariable String paper_id, @RequestBody TitleParam titleParam) {
         // todo:
-        paperMapper.modifyTitleById(Integer.parseInt(paper_id),titleParam.getTitle());
+        String owner = SecurityUtils.getSubject().getPrincipal().toString();
         Paper paper = paperMapper.selectPaperById(Integer.parseInt(paper_id));
+        if(!owner.equals(paper.getOwner())){
+            return ReturnValue.generate("800","无权限",null);
+        }
+        if(!ObjectUtils.isEmpty(titleParam.getTitle()))
+            paperMapper.modifyTitleById(Integer.parseInt(paper_id),titleParam.getTitle());
+        if(!ObjectUtils.isEmpty(titleParam.getAuthors()))
+            paperMapper.modifyAuthorsById(Integer.parseInt(paper_id),titleParam.getAuthors());
+        if(!ObjectUtils.isEmpty(titleParam.getPublisher()))
+            paperMapper.modifyPublisherById(Integer.parseInt(paper_id),titleParam.getPublisher());
+        if(titleParam.getYear() != 0)
+            paperMapper.modifyYearById(Integer.parseInt(paper_id),titleParam.getYear());
+        paper = paperMapper.selectPaperById(Integer.parseInt(paper_id));
         return ReturnValue.generateSuccess(new PaperParamOutput(paper));
     }
 
     @RequestMapping(value = "/api/paper/{paper_id}", method = RequestMethod.DELETE)
     public ReturnValue deletePaper(@PathVariable String paper_id){
-        paperMapper.deletePaperById(Integer.parseInt(paper_id));
+        String owner = SecurityUtils.getSubject().getPrincipal().toString();
         Paper paper = paperMapper.selectPaperById(Integer.parseInt(paper_id));
+        if(!owner.equals(paper.getOwner())){
+            return ReturnValue.generate("800","无权限",null);
+        }
+        paperMapper.deletePaperById(Integer.parseInt(paper_id));
+        //paper = paperMapper.selectPaperById(Integer.parseInt(paper_id));
         PaperParamOutput paperParamOutput = null;
         return ReturnValue.generateSuccess(paperParamOutput);
     }
 
     @PostMapping(value = "/api/paper/{paper_id}/pdf")
-    public  ReturnValue uploadPaperPDF(@PathVariable String paper_id, @RequestBody FileParam fileParam){
-        String uploadDir = "~/paper-management/src/PDF";
+    public  ReturnValue uploadPaperPDF(@PathVariable String paper_id, @RequestParam("file") MultipartFile file){
+        String owner = SecurityUtils.getSubject().getPrincipal().toString();
+        Paper paper = paperMapper.selectPaperById(Integer.parseInt(paper_id));
+        if(!owner.equals(paper.getOwner())){
+            return ReturnValue.generate("800","无权限",null);
+        }
+        String uploadDir = "/Users/lonelyslime/paper-management/src/PDF/";
         String fileName = UUID.randomUUID().toString() + ".pdf";
+        String pathname = uploadDir + fileName;
         try {
-            Path filePath = Paths.get(uploadDir, fileName);
-            Files.copy(fileParam.getFile().getInputStream(), filePath, StandardCopyOption.REPLACE_EXISTING);
-            paperMapper.releaseSourceByID(uploadDir + fileName,Integer.parseInt(paper_id));
+            File file1 = new File(pathname);
+            System.out.println("文件路径：%s".formatted(pathname));
+            file.transferTo(file1);
+            //Files.copy(file.getInputStream(), filePath, StandardCopyOption.REPLACE_EXISTING);
+            paperMapper.releaseSourceByID(pathname,Integer.parseInt(paper_id));
         } catch (IOException e) {
             throw new BizException("Failed to store file");
         }
-        return ReturnValue.generateSuccess(new SourceParam(uploadDir + fileName));
+        return ReturnValue.generateSuccess(new SourceParam(pathname));
+    }
+
+    @GetMapping(value = "/api/paper/{paper_id}/pdf/download")
+    @ResponseBody
+    public ResponseEntity<Resource> downloadPaper(@PathVariable String paper_id){
+        String source = paperMapper.selectPaperById(Integer.parseInt(paper_id)).getSource();
+        if(ObjectUtils.isEmpty(source)){
+            return null;
+        }
+        Resource resource = new FileSystemResource(source);
+        String filename = "paper.pdf";
+        return ResponseEntity.ok()
+                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + filename + "\"")
+                .body(resource);
+    }
+
+    @GetMapping(value = "/api/paper/{paper_id}/pdf/preview")
+    public ResponseEntity<byte[]> previewPaper(@PathVariable String paper_id) throws IOException {
+        String source = paperMapper.selectPaperById(Integer.parseInt(paper_id)).getSource();
+        if(ObjectUtils.isEmpty(source)){
+            return null;
+        }
+        Path path = Paths.get(source);
+        byte[] fileContent = Files.readAllBytes(path);
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_PDF);
+        return ResponseEntity.ok()
+                .headers(headers)
+                .body(fileContent);
+
     }
 }
